Factor out NYSDOT and MCNY logic
