+++ src/Data/Traffic/ONGOV.cpp
@@ -1,0 +1,8 @@
+#include "ONGOV.h"
+#include <string>
+
+namespace Traffic {
+namespace ONGOV {
+  extern const std::string EVENTS_URL{ "https://911events.ongov.net/CADInet/app/events.jsp" };
+}
+}
